refactor: in settings ui (#51)
