+++ app/src/main/java/io/kokoichi/sample/sakamichiapp/presentation/setting/SettingScreen.kt
@@ -10,4 +10,5 @@ sealed class SettingScreen(val route: String) {
     object ClearCacheScreen : SettingScreen("clear_cache")
     object ReportIssueScreen : SettingScreen("report_issue")
     object SetThemeScreen : SettingScreen("set_theme")
+    object ShareAppScreen: SettingScreen("share_app")
 }
