+++ app/src/main/java/io/kokoichi/sample/sakamichiapp/presentation/setting/SettingTopScreen.kt
@@ -1,16 +1,16 @@
 package io.kokoichi.sample.sakamichiapp.presentation.setting
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.*
+import androidx.compose.material.Divider
+import androidx.compose.material.Icon
+import androidx.compose.material.Text
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.KeyboardArrowRight
-import androidx.compose.runtime.*
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
@@ -20,17 +20,14 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import io.kokoichi.sample.sakamichiapp.BuildConfig
 import io.kokoichi.sample.sakamichiapp.R
+import io.kokoichi.sample.sakamichiapp.presentation.setting.components.VersionInfo
+import io.kokoichi.sample.sakamichiapp.presentation.ui.theme.IconSizeMedium
+import io.kokoichi.sample.sakamichiapp.presentation.ui.theme.SpaceSmall
+import io.kokoichi.sample.sakamichiapp.presentation.ui.theme.SpaceTiny
 import io.kokoichi.sample.sakamichiapp.presentation.ui.theme.Typography
-import io.kokoichi.sample.sakamichiapp.presentation.util.Constants
-import io.kokoichi.sample.sakamichiapp.presentation.util.DataStoreManager
+import io.kokoichi.sample.sakamichiapp.presentation.util.Constants.BottomPadding
 import io.kokoichi.sample.sakamichiapp.presentation.util.TestTags
-import io.kokoichi.sample.sakamichiapp.presentation.util.getMilliSecFromLocalTime
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.async
-import kotlinx.coroutines.launch
-import java.time.LocalTime
 
 @Composable
 fun SettingTopScreen(
@@ -41,7 +38,8 @@ fun SettingTopScreen(
 ) {
     Column(
         modifier = Modifier
-            .fillMaxSize(),
+            .fillMaxSize()
+            .padding(bottom = BottomPadding),
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
         TopBar()
@@ -86,12 +84,6 @@ fun SettingTopScreen(
         ) {
             viewModel.writeIsDevTrue(context)
         }
-        Spacer(
-            modifier = Modifier
-                .height(40.dp)
-                .fillMaxWidth()
-                .background(uiState.themeType.subColor)
-        )
     }
 }
 
@@ -99,7 +91,7 @@ fun SettingTopScreen(
 fun TopBar() {
     Text(
         modifier = Modifier
-            .padding(top = 10.dp, bottom = 5.dp)
+            .padding(top = SpaceSmall, bottom = SpaceTiny)
             .testTag(TestTags.SETTING_TITLE),
         text = stringResource(R.string.setting_screen_title),
         style = Typography.h5,
@@ -118,152 +110,23 @@ fun SettingRow(
             .clickable {
                 onclick()
             }
-            .padding(10.dp)
+            .padding(SpaceSmall)
             .testTag(text),
         contentAlignment = Alignment.CenterEnd,
     ) {
         Text(
             modifier = Modifier
-                .fillMaxWidth(),
+                .fillMaxWidth()
+                .padding(SpaceSmall),
             text = text,
             style = Typography.body2,
         )
         Icon(
             imageVector = Icons.Default.KeyboardArrowRight,
-            contentDescription = "right arrow",
+            contentDescription = stringResource(R.string.right_arrow),
             tint = Color.LightGray,
             modifier = Modifier
-                .size(30.dp)
-        )
-    }
-}
-
-@Composable
-fun VersionInfo(
-    borderColor: Color,
-    onIsDevChanged: () -> Unit,
-) {
-
-    val context = LocalContext.current
-    // For click timing
-    var touchNum by remember { mutableStateOf(0) }
-    var lastTouchedTime by remember { mutableStateOf(0) }
-
-    // For snackBar
-    val snackbarCoroutineScope = rememberCoroutineScope()
-    val snackbarHostState = remember { SnackbarHostState() }
-    var snackBarJob: Job? = null
-    val thresholdMillSec = 1000
-    val needTap = 7
-
-    var isDeveloper by remember { mutableStateOf(false) }
-    LaunchedEffect(Unit) {
-        val storeVal = async {
-            DataStoreManager.readBoolean(context, DataStoreManager.KEY_IS_DEVELOPER)
-        }
-        isDeveloper = storeVal.await()
-    }
-
-    val successMessage = stringResource(R.string.developer_success_snack_bar_text)
-    val onClick = when (isDeveloper) {
-        false -> {
-            {
-                val current = getMilliSecFromLocalTime(LocalTime.now())
-                if (current - lastTouchedTime < thresholdMillSec) {
-                    touchNum += 1
-                } else {
-                    touchNum = 1
-                }
-                lastTouchedTime = current
-                // If the count is near the needTapNum, show snack bar.
-                if (touchNum > 2) {
-                    snackBarJob?.cancel()
-                    snackBarJob = snackbarCoroutineScope.launch {
-                        snackbarHostState.showSnackbar("You are now ${needTap - touchNum} steps away from being a developer.")
-                    }
-                }
-                // If the count is over the needTapNum, show snack bar with special message.
-                if (touchNum >= needTap) {
-                    snackBarJob?.cancel()
-                    snackBarJob = snackbarCoroutineScope.launch {
-                        onIsDevChanged()
-                        snackbarHostState.showSnackbar(successMessage)
-                    }
-                }
-            }
-        }
-        true -> {
-            {}
-        }
-    }
-
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clickable {
-                onClick()
-            }
-            .padding(15.dp)
-            .testTag(TestTags.SETTING_VERSION),
-        contentAlignment = Alignment.Center,
-    ) {
-        Text(
-            modifier = Modifier
-                .fillMaxWidth(),
-            text = "バージョン情報 ${BuildConfig.VERSION_NAME}",
-            style = Typography.body2,
-            textAlign = TextAlign.Center,
-        )
-    }
-
-    SnackbarSetting(
-        snackbarHostState = snackbarHostState,
-        borderColor = borderColor,
-    )
-}
-
-@Composable
-fun SnackbarSetting(
-    snackbarHostState: SnackbarHostState,
-    borderColor: Color,
-) {
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .padding(Constants.BottomBarPadding)
-    ) {
-        SnackbarHost(
-            modifier = Modifier.align(Alignment.BottomCenter),
-            hostState = snackbarHostState,
-            snackbar = { snackbarData: SnackbarData ->
-                Card(
-                    shape = RoundedCornerShape(8.dp),
-                    modifier = Modifier
-                        .padding(16.dp)
-                        .wrapContentSize()
-                        .border(
-                            width = 1.dp,
-                            color = borderColor,
-                            shape = RoundedCornerShape(8.dp)
-                        )
-                ) {
-                    Column(
-                        modifier = Modifier.padding(8.dp),
-                        verticalArrangement = Arrangement.spacedBy(4.dp),
-                        horizontalAlignment = Alignment.CenterHorizontally
-                    ) {
-                        Text(
-                            modifier = Modifier
-                                .padding(5.dp)
-                                .testTag(TestTags.SNACK_BAR_TEXT),
-                            text = snackbarData.message,
-                            style = MaterialTheme.typography.body2.copy(
-                                color = Color.DarkGray,
-                            ),
-                        )
-                    }
-                }
-            }
+                .size(IconSizeMedium)
         )
     }
 }
