+++ app/src/main/java/io/kokoichi/sample/sakamichiapp/presentation/ui/theme/Dimensions.kt
@@ -7,3 +7,7 @@ val SpaceSmall = 8.dp
 val SpaceMedium = 16.dp
 val SpaceLarge = 24.dp
 val SpaceHuge = 32.dp
+
+val IconSizeSmall = 15.dp
+val IconSizeMedium = 25.dp
+val IconSizeLarge = 35.dp
